sf: Fix issue with dequeueBuffer when multiple display present
swapBuffers is called even for device composition when a flip client
target request is present. dequeueBuffer call is skipped
as there is no client composition. This is fine for single display
but if multiple display i.e. vds is present, this will be an issue
as current surface will not be switched and swapBuffers will fail

When flip client target request is present, make sure buffer is
dequeued before swapBuffers.

Bug: 151698217
Test: atest libsurfaceflinger_unittest
Change-Id: I933cbae2f09f81eef6555b1bb1e5991d2c450930
Merged-In: Ia8a1470affb2596b27986cc4153417f48cf4ed1c
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>