SF: Update forced composition color space along with color mode
Currently, a forced composition color space (ColorMode) can be set by
the device via a persistent property. However, this property is read
only during init, while some DisplayColorSetting values, which can
change during runtime, require a specific ColorMode to be set
simultaneously.

This change adds the ability to specify a ColorMode in the same binder
transaction that is used to set DisplayColorSetting. If a ColorMode is
unspecified during the transaction, the current forced ColorMode is
unchanged.

Bug: 137140317
Bug: 137053654
Change-Id: I35bb292fbe70f944f509f7d34ef965adb66ca059
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>