Use getLayerStack() instead of layerStack for input info
When there's no displayId, the layerStack should be used. The current
code was using mDrawingState.layerStack which is incorrect because only
the layerStack for the root of the tree is correct. Instead, use
getLayerStack() which traverses the hierarchy to find the root's
layerStack.

Test: Uses correct layerStack instead of 0 for input with unset displayIds
Fixes: 140312942
Change-Id: Ic8a06c562e970b0f328c539edfa51e206734934f
Merged-in: Ic8a06c562e970b0f328c539edfa51e206734934f
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>