Swap video frame rotation direction.
The DISPLAY_ORIENTATION_* values indicate the physical rotation of the
display, and not the rotation of the graphics surface (see
android.view.Display#getRotation()). To compensate for a physical
rotation in one direction, the video frame needs to be rotated in the
opposite direction.

Fixes: 150382187
Test: atest libinput_tests inputflinger_tests
Change-Id: Iddd943c35084f7032a8e60273e5e63b59ad10d1c
Merged-In: Iddd943c35084f7032a8e60273e5e63b59ad10d1c
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>